PS-397 Separate the code out
